requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# Telegram Bot
python-telegram-bot==20.7

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# Cache
redis==5.0.1

# HTTP Client
httpx==0.26.0

# Scheduling
apscheduler==3.10.4

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Utilities
python-dotenv==1.0.0
numpy==1.26.3
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
